iproto: connection shutdown
Connection shutdown flow:
    1. See in tx-thread if connection support graceful shutdown
    2.1. If connection support graceful shutdown, then send
         IPROTO_SHUTDOWN to remote client
    2.2 If connection does not support graceful shutdown, then read data
        from input socket until iproto packet boundary or until
        msg_max_limit is reached and shutdown(con, SHUT_RD).
    3. Close connection when connection becomes shutdown ready.

During the standart connection lifecycle connection can be closed. If it
is happend during the shutdown process, then it leads to UB, because at
the end of connection_close process connection is freed but part of
iproto_connection structure is used in shutdown process. Added extra
"return" in iproto_connection_try_to_start_destroy to avoid it.

Part of #5924