session: introduce `graceful_shutdown_required` session setting
New `graceful_shutdown_required` session setting  is designed
to separate sessions that support and do not support graceful
shutdown in iproto. By default, the value of this setting is
`false`. If the user needs graceful shutdown support in iproto,
he should set this setting value to `true`.
Part of #5924

@TarantoolBot document
Title: introduce `graceful_shutdown_required` session setting
New `graceful_shutdown_required` session setting  is designed
to separate sessions that support and do not support graceful
shutdown in iproto. By default, the value of this setting is
`false`. If the user needs graceful shutdown support in iproto,
he should set this setting value to `true`.