vinyl: fix race between compaction and gc of dropped LSM
This is a backport of the master commit
29e2931c66c5caf38138ce281bd558ff25984e38.

The difference between the current master and 1.10 is that the former
doesn't exclude dropped LSM trees from dump and compaction while the
latter does (see d7387ec99083e2132ac1d1a0889ccbf75b498dc5). So we can't
just cherry-pick the fix. Instead we patch vy_task_compaction_abort so
that it doesn't try to discard the newly written run in case the index
was dropped during compaction.

Closes #5436