txm: remove redundant _story suffix from memtx_tx_story_link
Once there were two type of links of a story: with a clean tuple
and with other story, that's why there was two similar functions
that were used for linking, with differen suffic in names. After
some refactoring the linkage with tuples was removed, so now
a story can only be linked with another story. The time has come
to remove meaningless suffix too.

Part of #6132