txm: allow only one TX at the time handling DDL
Until transactional DDL is implemented it makes no sense to allow
several transactions process DDL operations: there's no transactional
caches storing metadata of object, so it extremely easy to broke MVCC so
that one TX sees artifacts of another TX even if they are both open.
So let's allow only one TX at the time modify any data-dictionary.
To achieve this we store ID of the transaction that processes DDL
operation first; on TX commit/rollback we reset that ID. In case any
other TX attempts to handle DDL - we raise an error. Furthermore, we
have to store schema version in TX while it is started: transaction can
start long ago and some other TX may begin and commit DDL changes.
To avoid phantom reads, we check that schema version of TX to be
committed is newer than it was at the start.

Closes #5998