github-ci: pull all available tags
Found that Github actions/checkout@v2, when Github Actions triggers
a job upon the tag push, it makes the following calls [1]:

  git fetch origin +refs/heads/*:refs/remotes/origin/* +refs/tags/*:refs/tags/*
  git fetch origin +$HASH:refs/tags/$TAG

While the first call is fine, the second one spoils the tag and makes
it non-annotated.

Later, Tarantool calls git describe without --tags flag, and the
spoiled tag is ignored. This patch handles it by fetching all the
tags once again explicitly.

Also explained the next step, where we remove the tag. The Github
Actions may runs twice the same job on branch push and on new tag
if it was created on this push. To avoid of fails on saving the
commonly named packages, we drop a tag that points to a current
commit (if any) on a job triggered by pushing to a branch (it will
produce the package based on previous tag) and leave the new tag
in job triggered by tag.

To check this change was used commands to create dump commit and tag
the new version. Also made twice push with extra flag for tags:

  git commit -m"Dump commit" --allow-empty
  git tag -a 2.4.8 -m ''
  git push && git push --tags

In this way would be run 2 jobs for each package:

  1. Build package on testing branch. It will produce package with
     name 2.4.7.x (because of the exception mentioned above).
  2. Build package on tag 2.4.8. It will produce package with name
     2.4.8.x.

Closes tarantool/tarantool-qa#119

[1]: https://github.com/actions/checkout/issues/290

(cherry picked from commit 1c76fa519cb2db7e61c77e80370abb3abecfe02b)