build/curl: disable ZSTD content encoding support
I'm not against ZSTD content encoding.

The problem I want to solve is the following. I'll add a unit test in a
future commit:

 | add_executable(serializer.test serializer.c)
 | target_link_libraries(serializer.test unit box ${LUAJIT_LIBRARIES})

Linking of the test fails if curl's configure script detects ZSTD
presence and enables it. It is irrelevant for building of libcurl using
CMake (since 2.6.0-196-g2b0760192).

ZSTD content encoding support was added in curl-7.72.0, we pull it in
1.10.9-124-ged90a8c22 ('security: update libcurl from 7.71.1 to
7.76.0'). The key problem is that curl's configure script (unlike CMake)
performs autodetection of present libraries by default.

If we'll want to enable this content encoding, it is better to do so
deliberately: with explicit linking against our built-in ZSTD (now the
choice depends on presence of the library in the system) and correct
dependencies in our CMake scripts (correct CURL_LIBRARIES).

Part of #3228