txm: use index itself instead of index_id
There was a serious problem in txm: index_id from struct index was
used as an index in some arrays (for example in array of links in
stories). As a result, if a user had created an index specifying ID
that is not sequential, the array access would have been out of
range which could lead to segfault.

This patch makes use of indexes directly, and when it comes to
array aceess, a dense_id is used, which fits perfectly for that.

As a part of #5515 this patch makes the cases in it at least stable.

Part of #5515

(cherry picked from commit 8bd94bc095a1d031daf353bfd877c5e2811cc770)