lua: fix space:fselect() output with multibyte characters
Currently, we use len() to find length of string and then we use this value to calculate width of column. The problem is len() returns size of string in bytes, not number of characters in string. That is why alignment of space:fselect() output is being broken by multibyte characters. To avoid this, use utf8.len() instead of len().

The alignment will be still broken, if there are characters with variable length in space.

Closes #6165