net.box: drop conn:timeout()
conn:timeout(timeout) writes (now + timeout) to conn._deadlines table
keyed by fiber.self(). The deadline will then be used by all conn
methods unless specified explicitly in the options argument.

Usage of fiber.self() degrades net.box performance by about 15% in case
timeout is not specified, because it is a C call and so interferes with
JIT. See the ticket for the test details.

Since the feature is unusable and was deprecated more than four years
ago (in 1.7.4) in favor of specifying timeout in the options argument
(see commit 58da206c33197 ("net.box: add { timeout = } option to
requests")), it should be fine to drop it.

Closes #6242

@TarantoolBot document
Title: Delete documentation entry for net.box conn:timeout()

conn:timeout() was deprecated more than four years ago (in 1.7.4). Since
it results in performance degradation in case timeout is not specified
explicitly in request options, we decided to drop it. Please remove it
from the documentation as well:

https://www.tarantool.io/en/doc/latest/reference/reference_lua/net_box/#conn-timeout