txm: fix a dangerous misprint in gap handling
Read gap is a fact that a search by key in index found nothing.
In order to remember that fact a record is stored in the nearest
(to the right) tuple in that index; in case of no such tuple -
in special place in struct index.
In other words a gap record means that its transaction can be
vulnerable to inserts into that place in index. If something is
inserted into that gap - its transaction must fall into read view
or be aborted.

Further, if a tuple is inserted into that place of index, we have
to decide to which interval (or to both) that gap belongs. If its
owner transaction is vulnerable to both subintervals - the gap
record is splitted into two parts.

There's a special case when the search key is NULL, that stands
for index:select{} call (with empty table as an argument).
Obviously such a read is vulnerable to writes to any part of that
index, that means that such a gap record must always be splitted.

This patch remove an obvious misprint and makes gap to be splitted
in this case.

Part of #6206

(cherry picked from commit b1ee4e59b6450812fb68f34144d7ac7d639623bc)