backtrace: add RIP tracing and resolving API
For now only in-place backtrace is supported. Howewer, in some cases
it's good to have the ability to trace stack in a cheap way in order to
resolve names and addresses on-demand later.

This patch provides the ability to trace only RIPs of calls in a stack
and later resolve procedure names and addresses basing on these RIPs.
Due to the limited libunwind API for DARWIN this patch uses more
straightforward implementation of parent backtrace collection under OS X.
To resolve backtrace information in this case dladdr procedure is used.

Needed for: #4002