txm: track duplicated tuples
There was a bug when a transaction makes a wrong statement that is
aborted because of duplicate tuple in primary or secondary index.
The problem is that check of existing tuple is an implicit read
that has usual side effect.

This patch tracks that kind of reads like ordinal reads.

Part of #5999

(cherry picked from commit 2cafa623cd406dc153c59a26443cb7a9758dc425)