sql: fix quote() function
After this patch SQL built-in function QUOTE() will return the argument
in case DOUBLE value is given. If the argument is not number, string
representation of the argument will be returned.

Closes #6239

@TarantoolBot document
Title: QUOTE() and DOUBLE argument

After this patch function QUOTE() will return argument in case it
receives DOUBLE value as an argument. The same for all other numeric
types. In case it was given value of non-numeric type, it will return
string representation of the argument.