github-ci: provide commit id within manual
Git checkout :ref was added to workflows where packaging is located
This is done if someone wants to build packaging from their branch or
particular commit.

Long and Short commit sha can be added to the input. If nothing is added it
will checkout to HEAD branch and if branch/commit do not exist, failure will occur.

close #5416