box: allow to insert fraction-less doubles
It used to be forbidden to insert fraction-less numbers from Lua to
box.insert(). When we constructed tuples in Lua and passed them to
MessagePack this fraction-less doubles got converted to integers and
thus converted as MP_UINT/MP_INT, which is not expected for columns
of type DOUBLE. Then we got to use this awkward work-around

  space:insert({ffi.cast('double', 424.0})

We lift this limitation now, and allow integer values when we expect
doubles. So one could write simple:

  space:insert({424.0})

Such numbers will be accepted, and properly sorted.

For performance reasons, and to reduce a size of generated payload
we do not enforce MP_DOUBLE when we serialize to MessagePack, but rather
tolerate current MP_UINT/MP_INT. We rather modify decoder part to act
accordingly. Compare, for example, sizeof of MessagePack payload for 0
if MP_UINT - it will be 1 byte, while 9 bytes if MP_DOUBLE used.

Closes #5933