github-ci: support the new version format
The new release policy suggests to use X.Y.Z-alphaN, X.Y.Z-betaN,
X.Y.Z-rcN tags for pre-releases. See the policy document here:
https://github.com/tarantool/tarantool/discussions/6182

This commit enforces the new versioning for tarantool builds that're
made from CI.

Tarballs
--------

When we'll add the 2.10.0-beta1 tag, the tarball will be named so:

> tarantool-2.10.0-beta1.tar.gz

When we'll add one commit upward the tag, the tarball will be named so:

> tarantool-2.10.0-beta1.1.dev.tar.gz

(In fact, we don't deploy tarballs per push anymore, so it is just for
the same of completeness.)

For 2.10.0 release it'll be named as:

> tarantool-2.10.0.tar.gz

How to try: `make -f .gitlab.mk source`.

Packages
--------

When we'll add the 2.10.0-beta1 tag, the packages for Ubuntu Focal will
be named so:

> tarantool_2.10.0~beta1-1_amd64.deb
> tarantool-common_2.10.0~beta1-1_all.deb
> tarantool-dev_2.10.0~beta1-1_amd64.deb

The sources packaged together with the package will be named so:

> tarantool_2.10.0~beta1-1.debian.tar.xz
> tarantool_2.10.0~beta1-1.dsc
> tarantool_2.10.0~beta1.orig.tar.xz
> tarantool-2.10.0~beta1.tar.xz

When we'll add one commit upward the tag, those files will look so:

> tarantool_2.10.0~beta1.1.dev-1_amd64.deb
> tarantool-common_2.10.0~beta1.1.dev-1_all.deb
> tarantool-dev_2.10.0~beta1.1.dev-1_amd64.deb

> tarantool_2.10.0~beta1.1.dev-1.debian.tar.xz
> tarantool_2.10.0~beta1.1.dev-1.dsc
> tarantool_2.10.0~beta1.1.dev.orig.tar.xz
> tarantool-2.10.0~beta1.1.dev.tar.xz

(In fact, we don't deploy packages per push anymore, so it is just for
the same of completeness.)

For 2.10.0 release it'll be named so:

> tarantool_2.10.0.gcb3bdbf2a-1_amd64.deb
> tarantool-common_2.10.0.gcb3bdbf2a-1_all.deb
> tarantool-dev_2.10.0.gcb3bdbf2a-1_amd64.deb

> tarantool_2.10.0.gcb3bdbf2a-1.debian.tar.xz
> tarantool_2.10.0.gcb3bdbf2a-1.dsc
> tarantool_2.10.0.gcb3bdbf2a.orig.tar.xz
> tarantool-2.10.0.tar.xz

How to try: `OS=ubuntu DIST=focal make -f .gitlab.mk package`.

Fixes #6184