test: descrease instance file name length
First, several facts:

* The maximal unix socket file length is 108 symbols on Linux (in fact,
  107 for tarantool console socket, see #4634).
* We run testing during building of RPM packages.
* When tests are run, they use the build directory for temporary testing
  files and tarantool console unix socket files.
* Test-run uses an instance script name to form a unix socket name for
  tarantool console.
* We plan to change package version format and add .dev suffix for
  transient (non tagged) commits.
* The package version participates in the build directory name, when
  we're building an RPM package.

If we'll enable the new version format just now, we'll unable to listen
on the unix socket like the following (the length is 111):

```
/build/usr/src/debug/tarantool-2.9.0.353.dev/test/var/023_box/gh-3633-simple-tuple-size-increasing.socket-admin
```

So I changed the instance file name to make it shorter.

Of course, there are ways to solve the problem at all, this commit is
just workaround.

Part of #6184
Follows up #3633